requirements.txt
# dev requirements
aiohttp==3.8.4
black==23.3.0

# package requirements
apscheduler==3.10.1
python-dotenv==1.0.0
